Implement custom notifications
This commit adds a new type of notification (called "custom"), that allow developers
to tailor notifications for their application.

Custom notifications are defined by: a widget, a message, and a set of parameters. The
end-user will be shown the message, while the "View" button will take the uset to the
widget passing the parameters as props.