Fixed undefined state
Sometimes the variable `state.items` will go undefined, this commit
is a patch to this error to which I was **not able to find the
root source**.

Testing different widgets in the vscode extension, I noticed that
`state.items` will sometimes go `undefined`, stopping the Feed
from loading widgets. However, despite all of my efforts, I was
not able to trace when in the code `state.items` goes undefined.