Fix: On overview page display logged in user instead of the moderator account.
Fix: Await graphql call in Moderation indexer to avoid indexer failure if there is an error.

Fix: Moderating an account from the Needs Moderation page writes to an incorrect Social.index key due to undefined path being appended.

Fix: Remove  receipt_id column in the needs_moderation as it is causing duplicates.

Fix: Correct subquery join for reporter list in unmoderated_reports to pull in reporters for reported accounts.

Feature: Disable Moderation buttons on Needs Moderation interface when not logged in as a moderator.